Restructured the files and directories
