+++ backend/storage/models.go
@@ -1,4 +1,4 @@
-package main
+package storage
 
 import (
 	"time"
